Laboratorio3, Universidad De Costa Rica, Recinto Tacares

+++ Laboratorio3/src/Laboratorio3/Ingresos.java
@@ -8,7 +8,7 @@ package Laboratorio3;
  */
 public interface Ingresos {
 
-    public int calcularIngresos();
+    public double calcularIngresos();
